Make pandas an optional dependency
Importing pandas/numpy can be slow (it doubles the startup time of
./manage.py, and some people reports that pandas takes seconds to load),
and it's a pretty huge dependency for a functionality that is already in
the standard library. Making it an optional dependency makes it easy for
people to gain pandas speedup while not incurring the cost for those who
don't need it.